moved getQuestions
