+++ src/controllers/exams/getQuestions.controller.js
@@ -1,0 +1,50 @@
+import Question from "#models/question.model.js";
+import { ApiError } from "#utils/api-handler/error.js";
+import { ApiResponse } from "#utils/api-handler/response.js";
+import { ApiHandler } from "#utils/api-handler/handler.js";
+
+export const getQuestions = ApiHandler(async (req, res) => {
+  const exam_id = req.query.exam_id;
+  const page = parseInt(req.query.page) || 1;
+  const limit = parseInt(req.query.limit) || 10;
+  const offset = (page - 1) * limit;
+
+  if (!exam_id) {
+    throw new ApiError(400, "BAD_REQUEST");
+  }
+
+  const { rows, count: total } = await Question.findAndCountAll({
+    where: { exam_id: exam_id },
+    offset,
+    limit,
+    attributes: ["id", "question_text", "type", "metadata"],
+    order: [["created_at", "ASC"]],
+  });
+
+  const questions = rows.map((question) => {
+    const { id, question_text, type, metadata } = question;
+
+    const sanitizedMetadata = { ...metadata };
+    if (req.user.role === "STUDENT") {
+      delete sanitizedMetadata.correct_answers;
+    }
+
+    return {
+      id,
+      question_text,
+      type,
+      metadata: sanitizedMetadata,
+    };
+  });
+
+  // Send response
+  return res.status(200).json(
+    new ApiResponse("EXAMS_FETCHED", {
+      total,
+      page,
+      limit,
+      totalPages: Math.ceil(total / limit),
+      questions,
+    })
+  );
+});
